Working mapper 24 with audio! Mapper 26 will be done soon.

+++ src/audio/DMC.java
@@ -137,7 +137,7 @@ public class DMC extends Channel{
 		
 	}
 	@Override
-	public int getOutput(){
+	public double getOutput(){
 		return outputlevel;	
 	}
 	@Override
